Fix OSM queries

+++ backend/src/main/kotlin/io/github/lauzhack/backend/api/openStreetMap/OpenStreetMapResponse.kt
@@ -5,15 +5,15 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class OpenStreetMapResponse(
-    @SerialName("place_id") val placeId: Int,
+    @SerialName("place_id") val placeId: Long,
     @SerialName("licence") val licence: String,
     @SerialName("osm_type") val osmType: String,
-    @SerialName("osm_id") val osmId: Int,
+    @SerialName("osm_id") val osmId: Long,
     @SerialName("lat") val lat: String,
     @SerialName("lon") val lon: String,
     @SerialName("class") val classX: String,
     @SerialName("type") val type: String,
-    @SerialName("place_rank") val placeRank: Int,
+    @SerialName("place_rank") val placeRank: Long,
     @SerialName("importance") val importance: Double,
     @SerialName("addresstype") val addressType: String,
     @SerialName("name") val name: String,
